Refactor atom
