+++ LEGO1/omni/include/mxsoundmanager.h
@@ -2,7 +2,7 @@
 #define MXSOUNDMANAGER_H
 
 #include "decomp.h"
-#include "mxatomid.h"
+#include "mxatom.h"
 #include "mxaudiomanager.h"
 
 #include <dsound.h>
